learn_ai/requirements.txt
django
langchain_community
python-dotenv
streamlit
langchain
groq
faiss-cpu
huggingface_hub
bs4
sentence-transformers
pypdf
